ALmost Complete Linux Compatibility (Arch Linux with X11 Tested)
Fixed AppCache issue in JSX.
Fixed AppDir and AppConfig compatibility in JSX (Both are the same on Windows but not on Unix systems)
Added Target OS everywhere when necessary.
Plannification of the creation of multiple different modules to be called for each OS.
Added todo executable target (Working On It)
Added Clippy Usage for cleaner code.