Design and Delete Files :
Fixed Z-Index Issues with the Sidebar.
Added optional chaining that fixes infinite loop of errors.
Added better StopTorrent allowing the user to also delete the files that have been downloaded, for both horizontal and vertical slide.
Added delete_torrent and its api call to stop the torrent and delete the files that has been downloaded and/or pre-allocated.