refactor: Download page, main bootstrap, bittorrent+ddl api
Refactored a few things for better scalability and work down the line such as downloads page that is now finally not a huge mess of files.

The main file isn't 800 lines anymore and is now divided into a bootstrap.

I am aware of a bug caused by RA on the main file but everything run normally.

Switched from sending requests to aria2 to polling from events, way better, not done yet for the frontend part but should be by next patch.

Also optimized a few things.